+++ packages/brick_offline_first/lib/offline_first_with_rest.dart
@@ -32,6 +32,16 @@ abstract class OfflineFirstWithRestAdapter<_Model extends OfflineFirstWithRestMo
 /// compiler/analyzer.
 abstract class OfflineFirstWithRestRepository
     extends OfflineFirstRepository<OfflineFirstWithRestModel> {
+  /// If the response returned from the client is one of these error codes, the request
+  /// **will not** be removed from the queue. For example, if the result of a request produces a
+  /// 404 status code response (such as in a Tunnel not found exception), the request will
+  /// be reattempted. If [upsert] response matches this status code, it **will not** throw
+  /// an exception.
+  ///
+  /// Defaults to `[404, 501, 502, 503, 504]`.
+  @protected
+  final List<int> reattemptForStatusCodes;
+
   /// The type declaration is important here for the rare circumstances that
   /// require interfacting with [RestProvider]'s client directly.
   @override
@@ -56,9 +66,7 @@ abstract class OfflineFirstWithRestRepository
     bool autoHydrate,
     String loggerName,
     this.throwTunnelNotFoundExceptions = false,
-
-    /// Forwarded to [OfflineQueueHttpClient#reattemptForStatusCodes]
-    List<int> reattemptForStatusCodes,
+    this.reattemptForStatusCodes = const [404, 501, 502, 503, 504],
   })  : remoteProvider = restProvider,
         super(
           autoHydrate: autoHydrate,
@@ -144,6 +152,11 @@ abstract class OfflineFirstWithRestRepository
     } on RestException catch (e) {
       logger.warning('#upsert rest failure: $e');
       if (_ignoreTunnelException(e)) {
+        return instance;
+      }
+
+      // since we know we'll reattempt this request, an exception does not need to be reported
+      if (reattemptForStatusCodes.contains(e.response?.statusCode)) {
         return instance;
       }