+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_custom_offline_first_serdes.dart
@@ -20,7 +20,7 @@ Future<Map<String, dynamic>> _$CustomOfflineFirstSerdesToRest(
     OfflineFirstRepository repository}) async {
   return {
     'string': instance.string?.toRest(),
-    'strings': instance.strings?.map((Serializable c) => c?.toRest())
+    'strings': instance.strings?.map((Serializable c) => c?.toRest())?.toList()
   };
 }