+++ packages/brick_rest_generators/lib/src/rest_serialize.dart
@@ -42,9 +42,9 @@ class RestSerialize<_Model extends RestModel> extends RestSerdesGenerator<_Model
       // Iterable<enum>
       if (argTypeChecker.isEnum) {
         if (fieldAnnotation.enumAsString) {
-          return "$fieldValue?.map((e) => e.toString().split('.').last)";
+          return "$fieldValue?.map((e) => e.toString().split('.').last)?.toList()";
         } else {
-          return '$fieldValue?.map((e) => ${checker.argType.getDisplayString()}.values.indexOf(e))';
+          return '$fieldValue?.map((e) => ${checker.argType.getDisplayString()}.values.indexOf(e))?.toList()';
         }
       }