+++ packages/brick_rest_generators/CHANGELOG.md
@@ -1,1 +1,4 @@
 ## Unreleased
+
+* Fix a a JSON encode error in REST serialization. `.map` returns a `MappedListIterable` which `jsonEncode` cannot parse. It can parse `List<dynamic>`.
+* Fix REST's iterable enum serialization